Clarify management protobuf/gRPC API comments.
Make it more explicit how to use each field under which circumstances
and what the API expects from implementors (both client and server).

Note how the comments are formulated: They _tell_ users of the API,
how it is expected to be used and how to handle each field. A user
is not only a client (like ctl) but also a server (like management).
By defining the way the messages and fields are supposed to be used, in
particular whether they are required or not and their limitations, it is
made sure that all implementors use them the right way. If they adhere
to that "contract" in their implementation.

For example, if a field is meant to be "required", it means that 1)
the provider of that message _must_ provide this field and 2) the the
consumer _can_ rely on this field being present, no matter what.